requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.1
python-dotenv>=0.19.0
pyyaml>=6.0

# Trading APIs
yfinance>=0.1.74
ccxt>=3.0.0

# AI/ML
scikit-learn>=1.0.0
tensorflow>=2.8.0
torch>=1.11.0

# Data analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Database
sqlalchemy>=1.4.0

# Logging
loguru>=0.6.0

# Web framework (for API/dashboard)
fastapi>=0.75.0
uvicorn>=0.17.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0